refactor(schedules): tighten template picker payload plumbing
- Move serialization onto ScheduleTemplate.to_picker_dict so tests can
  target a public API instead of importing a leading-underscore helper.
- Restrict the template queryset with .only(PICKER_FIELDS) so the prompt
  TextField (can be large) isn't fetched on every list/form render.
- Collapse a redundant wrapper div in _template_gallery.html.
