+++ daiv/schedules/views.py
@@ -1,6 +1,5 @@
 import json
 import logging
-from typing import TYPE_CHECKING
 
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
@@ -22,9 +21,6 @@ from accounts.mixins import AdminRequiredMixin, BreadcrumbMixin
 from schedules.forms import ScheduledJobCreateForm, ScheduledJobUpdateForm, ScheduleTemplateForm
 from schedules.models import ScheduledJob, ScheduleTemplate
 
-if TYPE_CHECKING:
-    from collections.abc import Iterable
-
 logger = logging.getLogger("daiv.schedules")
 
 
@@ -43,26 +39,13 @@ def _subscriber_initial_json(schedule) -> str:
     return json.dumps(rows)
 
 
-def _template_picker_payload(templates: Iterable[ScheduleTemplate]) -> list[dict]:
-    """Serialize templates into the JSON shape the gallery drawer consumes.
+def _template_picker_payload() -> list[dict]:
+    """Build the gallery drawer's embedded JSON payload from all templates.
 
-    Deliberately excludes ``prompt`` — the gallery shows the description only;
-    the prompt flows through the server-side ``?template=<id>`` prefill path.
+    ``.only(*PICKER_FIELDS)`` keeps the ``prompt`` TextField (which can be
+    large) out of the query result — it is never rendered in the gallery.
     """
-    return [
-        {
-            "id": t.id,
-            "name": t.name,
-            "description": t.description,
-            "repo_id": t.repo_id,
-            "ref": t.ref,
-            "frequency_display": t.get_frequency_display(),
-            "frequency_summary": t.frequency_summary,
-            "notify_on_display": t.get_notify_on_display(),
-            "use_max": t.use_max,
-        }
-        for t in templates
-    ]
+    return [t.to_picker_dict() for t in ScheduleTemplate.objects.only(*ScheduleTemplate.PICKER_FIELDS)]
 
 
 class ScheduleListView(_ScheduleOwnerMixin, LoginRequiredMixin, ListView):
@@ -76,7 +59,7 @@ class ScheduleListView(_ScheduleOwnerMixin, LoginRequiredMixin, ListView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context["schedule_templates"] = _template_picker_payload(ScheduleTemplate.objects.all())
+        context["schedule_templates"] = _template_picker_payload()
         return context
 
 
@@ -113,7 +96,7 @@ class ScheduleCreateView(BreadcrumbMixin, _ScheduleOwnerMixin, SuccessMessageMix
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context["subscriber_initial_json"] = "[]"
-        context["schedule_templates"] = _template_picker_payload(ScheduleTemplate.objects.all())
+        context["schedule_templates"] = _template_picker_payload()
         tpl = self._get_template()
         context["selected_template_id"] = str(tpl.pk) if tpl is not None else ""
         return context
